notifications/notifications_service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
orjson==3.10.18
pydantic-settings==2.9.1
# Dev requirements
black==25.1.0
isort==6.0.1
flake8==7.2.0
aio-pika==9.5.5
httpx
circuitbreaker==2.1.3
pytz